2week

+++ yoonseyeon/2week/kgToPound.py
@@ -1,3 +1,3 @@
-kg=float(input("변환하고 싶은 kg을 입력하세요.: "))
+kg=int(input("변환하고 싶은 kg을 입력하세요.: "))
 pound=float(kg)*2.2041
 print(kg,"kg는", pound,"pound입니다.")
